fix: fold invocation bootstrap fixture into factory_build for pkg file-count
CI pkg-file-count limited pkg/service to 15 files; merge invocation_bootstrap_fixture.go into factory_build.go beside InvocationBootstrap wiring.
